% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkManova.R
\name{checkManova}
\alias{checkManova}
\title{MANOVA analysis of cluster}
\usage{
checkManova(clust)
}
\arguments{
\item{clust}{cluster object}
}
\value{
statistic of MANOVA
}
\description{
MANOVA analysis based on Pillai Statistic
}

