% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen1.R
\docType{data}
\name{data.gen1}
\alias{data.gen1}
\title{Data Generate 1}
\format{A data frame with 120 rows and 4 variables:
\describe{
  \item{V1}{Variable 1}
  \item{V2}{Variable 2}
  \item{V3}{Variable 3}
  \item{LABEL}{Labeling factor}
}}
\source{
generated randomly
}
\usage{
data.gen1
}
\description{
A dataset containing generated data for simulation
}
\keyword{datasets}

