% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen2.R
\docType{data}
\name{data.gen2}
\alias{data.gen2}
\title{Data Generate 2}
\format{A data frame with 40 rows and 4 variables:
\describe{
  \item{V1}{Variable 1}
  \item{V2}{Variable 2}
  \item{V3}{Variable 3}
  \item{LABEL}{Labeling factor}
}}
\source{
generated via \code{\link{MASS::mvrnorm}}
}
\usage{
data.gen2
}
\description{
A dataset containing generated data for simulation
}
\keyword{datasets}

