% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDataCluster.R
\name{managedata}
\alias{managedata}
\title{Preparing data for clustering.}
\usage{
managedata(var.choice)
}
\arguments{
\item{var.choice}{Chosen Variables of Dataset}
}
\value{
data.cluster Dataset with chosen variables
}
\description{
This function used to construct data for clustering from
dataset with chosen variables.
}
\details{
Don't use it from user.
}

