% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputPlugin.R
\name{pluginInput}
\alias{pluginInput}
\title{Input Plugin of Fuzzy Clustering on Rcmdr}
\usage{
pluginInput()
}
\description{
Graphical User Interface on Rcmdr Plugin.
This Plugin provide Interface to select variables of dataset that will be
used for Fuzzy Clustering, methods selection, and parameter specification
}

