% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result.GUI}
\alias{result.GUI}
\title{Result GUI}
\usage{
result.GUI(parent, cluster, valid, manov, method)
}
\arguments{
\item{parent}{parent window}

\item{cluster}{cluster object}

\item{valid}{validation index object}

\item{manov}{manova object}

\item{method}{method of clustering}
}
\description{
Result GUI
}

