% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softensemble.R
\name{soft.vote.ensemble}
\alias{soft.vote.ensemble}
\title{Soft Voting Cluster Ensemble}
\usage{
soft.vote.ensemble(data, seed, method = "FCM", K = 2, m = 2, gamma = 0,
  rho = rep(1, K), threshold = 10^-5, maxiter = 100)
}
\arguments{
\item{data}{data frame nxp}

\item{seed}{number of ensemble}

\item{method}{fuzzy clustering method that will be used ("FCM" or "GK")}

\item{K}{specific number of cluster (must be >1)}

\item{m}{fuzzifier / degree of fuzziness}

\item{gamma}{parameter of Gustafson Kessel Clustering}

\item{rho}{parameter of volume clustering in Gustafson Kessel Clustering}

\item{threshold}{threshold of convergence}

\item{max.iteration}{maximum iteration to convergence}
}
\value{
func.obj objective function that calculated.

U matrix n x K consist fuzzy membership matrix

V matrix K x p consist fuzzy centroid

D matrix n x K consist distance of data to centroid that calculated

Clust.desc cluster description (dataset with additional column of cluster label)

seeding list of random number that used as seeding
}
\description{
This function used to perform Soft Voting Cluster Ensemble
}
\details{
This function perform Soft Voting Cluster Ensemble algorithm
}

