% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biploting.R
\name{biploting}
\alias{biploting}
\title{Biploting Cluster Result}
\usage{
biploting(cluster)
}
\arguments{
\item{cluster}{a cluster object}
}
\value{
biplot a biplot
}
\description{
Biploting Cluster Result
}
\details{
Make Visualization Biplot from cluster analysis result
}
\examples{
library(RcmdrPlugin.FuzzyClust)
fuzzy.CM(X=iris[,1:4],K = 3,m = 2,RandomNumber = 1234)->cl
biploting(cl)
}

