% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkManova.R
\name{checkManova}
\alias{checkManova}
\title{MANOVA analysis of cluster}
\usage{
checkManova(clust)
}
\arguments{
\item{clust}{cluster object}
}
\value{
statistic of MANOVA
}
\description{
MANOVA analysis based on Pillai Statistic
}
\examples{
library(RcmdrPlugin.FuzzyClust)
fuzzy.CM(X=iris[,1:4],K = 3,m = 2,RandomNumber = 1234)->cl
checkManova(cl)
}

