% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen3.R
\docType{data}
\name{data.gen3}
\alias{data.gen3}
\title{Data Generate 3}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{V1}{Variable 1}
  \item{V2}{Variable 2}
  \item{LABEL}{Labeling factor}
}}
\source{
generated via \code{\link{MASS::mvrnorm}}
}
\usage{
data.gen3
}
\description{
A dataset containing generated data for simulation
}
\keyword{datasets}

