% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplotting.R
\name{radar.plotting}
\alias{radar.plotting}
\title{Radar Ploting Cluster Result}
\usage{
radar.plotting(cluster)
}
\arguments{
\item{cluster}{a cluster object}
}
\value{
radarplot a radarplot
}
\description{
Radar Ploting Cluster Result
}
\details{
Make Visualization Radar Ploting from
}
\examples{
library(RcmdrPlugin.FuzzyClust)
fuzzy.CM(X=iris[,1:4],K = 3,m = 2,RandomNumber = 1234)->cl
checkManova(cl)
}

