% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationindex.R
\name{validation.index}
\alias{validation.index}
\title{Validation Index of Fuzzy Clustering}
\usage{
validation.index(cluster)
}
\arguments{
\item{cluster}{Cluster Result from Fuzzy Clustering}
}
\value{
XB.index Xie Beni index

K.index Kwon index

MPC.index Modified Partition Coeeficient

CE.index Classification Entropy
}
\description{
Validation Index of Fuzzy Clustering
}
\details{
This function provide validation index that calculated from fuzzy clustering
result. There are 4 index that calculated, Xie Beni, Kwon, MPC, and CE index. Both three indexes
calculated from fuzzy membership and data point.

Xie Beni index calculated compactness and separation of clustering.

Kwon index extended Xie Beni index to eliminate its tendency to monotonically decrease when number of cluster approach the number of data point.

The best cluster result can be decided with minimum value of index.
}

